添加record POJO
